Rename ScheduleMainPageView to BookmarksView to better represent its purpose. Rename MainAppView to AppView to better represent its purpose. Rename MainAppViewModel to ParentViewModel to better represent its purpose. Move buttons to General/Buttons. Move AppView to root of /App
