+++ tumble-ios/App/ViewModels/BookmarksView-ViewModel.swift
@@ -8,32 +8,32 @@
 import Foundation
 import SwiftUI
 
-enum ScheduleViewType: String {
+enum BookmarksViewType: String {
     case list = "List"
     case calendar = "Calendar"
     
-    static let allValues: [ScheduleViewType] = [list, calendar]
+    static let allValues: [BookmarksViewType] = [list, calendar]
 }
 
-enum ScheduleMainPageStatus {
+enum BookmarksViewStatus {
     case loading
     case loaded
     case uninitialized
     case error
 }
 
-extension ScheduleMainPageView {
-    @MainActor final class ScheduleMainPageViewModel: ObservableObject {
+extension BookmarksView {
+    @MainActor final class BookmarksViewModel: ObservableObject {
         
         @Inject var scheduleService: ScheduleService
         @Inject var preferenceService: PreferenceService
         @Inject var courseColorService: CourseColorService
         
-        @Published var scheduleViewTypes: [ScheduleViewType] = ScheduleViewType.allValues
-        @Published var status: ScheduleMainPageStatus = .loading
+        @Published var scheduleViewTypes: [BookmarksViewType] = BookmarksViewType.allValues
+        @Published var status: BookmarksViewStatus = .loading
         @Published var days: [DayUiModel] = []
         @Published var courseColors: CourseAndColorDict = [:]
-        @Published var defaultViewType: ScheduleViewType
+        @Published var defaultViewType: BookmarksViewType
         
         init () {
             self.defaultViewType = .list
@@ -60,7 +60,7 @@ extension ScheduleMainPageView {
             }
         }
         
-        func onChangeViewType(viewType: ScheduleViewType) -> Void {
+        func onChangeViewType(viewType: BookmarksViewType) -> Void {
             let viewTypeIndex: Int = scheduleViewTypes.firstIndex(of: viewType)!
             preferenceService.setViewType(viewType: viewTypeIndex)
         }
